requirements.txt
streamlit
pandas
openpyxl
pillow
opencv-python-headless
pytesseract
numpy
